+++ tests/targets/fn_module.py
@@ -2,5 +2,12 @@ def bar():
     return "bar"
 
 
+def bar2():
+    return "bar2"
+
+
 def deprecated():
     return "deprecated"
+
+
+__all__ = ["bar", "bar2", "deprecated"]