+++ docs/release-notes.md
@@ -1,5 +1,19 @@
 # Release notes
 
+## Version 0.0.4
+
+### Added
+
+- `find_missing` test method.
+- `getter` attribute saving the injected getter.
+- `absolutify_import` helper.
+- Add pre-commit.
+
+### Changed
+
+- Rename typo `settings_preload_name` to `settings_preloads_name`.
+- Fix relative imports.
+
 ## Version 0.0.3
 
 ### Added